split models from coordinator
